ai-services/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
openai==0.28.1
replicate==0.12.0
python-dotenv==1.0.0
gunicorn==21.2.0
mediapipe==0.10.5
face-recognition==1.3.0
dlib==19.24.2